requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-active-link==0.1.8
django-ckeditor==6.2.0
django-crispy==1.8.1
django-js-asset==2.0.0
environ==1.0
environs==9.5.0
gunicorn==20.1.0
marshmallow==3.14.1
Pillow==9.0.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
sqlparse==0.4.2
whitenoise==6.0.0
